# docs/rfq_manager_ms_openapi_v1.yaml
openapi: 3.0.3
info:
  title: rfq_manager_ms — RFQ Lifecycle Management API
  description: |
    Interface contract for the RFQ Manager microservice (V1).
    Front-to-back design: 9 UI screens → 47 user actions → 6 resources → 28 endpoints → 9 tables.
  version: "1.0.0"
  contact:
    name: RFQ Manager Team
servers:
  - url: /rfq-manager/v1
    description: Base path — convention /{ms_name}/{version}
...
security:
  - bearerAuth: []

# ═══════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════
paths:
...
  /workflows/{workflowId}:
    get:
      tags: [Workflow]
      summary: "#8 — Get workflow detail"
      description: Returns workflow + ordered stages[] with id, name, order, default_team, planned_duration_days.
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
      responses:
        '200':
          description: Workflow with stages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Workflow]
      summary: "#9 — Update workflow"
      description: "Updatable: name, description, is_active, is_default."
      operationId: updateWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowUpdateRequest'
      responses:
        '200':
          description: Workflow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  # ──── RFQ_STAGE ──────────────────────────────────
...
  /rfqs/{rfqId}/stages/{stageId}/subtasks:
    post:
      tags: [Subtask]
      summary: "#16 — Create subtask"
      description: "Auto-sets: progress = 0, status = Open."
      operationId: createSubtask
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubtaskCreateRequest'
      responses:
        '201':
          description: Subtask created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subtask'

    get:
      tags: [Subtask]
      summary: "#17 — List subtasks"
      description: Returns active subtasks (WHERE deleted_at IS NULL).
      operationId: listSubtasks
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
      responses:
        '200':
          description: Subtask list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Subtask'
  /rfqs/{rfqId}/stages/{stageId}/subtasks/{subtaskId}:
    patch:
      tags: [Subtask]
      summary: "#18 — Update subtask"
      description: |
        Updatable: name, assigned_to, due_date, progress, status.
        When all subtasks = 100% → stage progress auto-updates.
      operationId: updateSubtask
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
        - $ref: '#/components/parameters/SubtaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubtaskUpdateRequest'
      responses:
        '200':
          description: Subtask updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subtask'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Subtask]
      summary: "#19 — Delete subtask"
      description: Soft delete — sets deleted_at = NOW(). Keeps for audit.
      operationId: deleteSubtask
      parameters:
        - $ref: '#/components/parameters/RfqIdParam'
        - $ref: '#/components/parameters/StageIdParam'
        - $ref: '#/components/parameters/SubtaskIdParam'
      responses:
        '204':
          description: Subtask soft-deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ──── FILE (shortcut) ────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from rfq_iam_ms

  # ──── PARAMETERS ─────────────────────────────────
  parameters:
    RfqIdParam:
      name: rfqId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    WorkflowIdParam:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    StageIdParam:
      name: stageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SubtaskIdParam:
      name: subtaskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Not found"
        message:
          type: string
          example: "RFQ with id 'xxx' does not exist"

    # ── RFQ ──
...
    WorkflowSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "GHI long workflow"
        code:
          type: string
          example: "GHI-LONG"
        stage_count:
          type: integer
          example: 11
        is_active:
          type: boolean
        is_default:
          type: boolean
    WorkflowDetail:
      allOf:
        - $ref: '#/components/schemas/WorkflowSummary'
        - type: object
          properties:
            description:
              type: string
            stages:
              type: array
              items:
                $ref: '#/components/schemas/StageTemplate'
    WorkflowUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        is_default:
          type: boolean
    StageTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "RFQ received"
        order:
          type: integer
          example: 1
        default_team:
          type: string
          example: "Sales / BD"
        planned_duration_days:
          type: integer
          example: 2

    # ── RFQ_STAGE ──
...
    SubtaskCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Finish vessel plates BOQ"
        assigned_to:
          type: string
          example: "Mechanical"
        due_date:
          type: string
          format: date
    SubtaskUpdateRequest:
      type: object
      properties:
        name:
          type: string
        assigned_to:
          type: string
        due_date:
          type: string
          format: date
        progress:
          type: integer
        status:
          type: string
          enum: [Open, In progress, Done]
    Subtask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        assigned_to:
          type: string
          nullable: true
        due_date:
          type: string
          format: date
          nullable: true
        progress:
          type: integer
        status:
          type: string
          enum: [Open, In progress, Done]
        created_at:
          type: string
          format: date-time

    # ── REMINDER ──
...